refactor(sh): move CompletedCommand from errors.py to shell_command.py
CompletedCommand is the success-path return type of run_command, not an
error. It lived in errors.py only to avoid an import cycle with the error
classes that reference it via the .result attribute. With
`from __future__ import annotations` and a TYPE_CHECKING import in
errors.py, the cycle dissolves and the type can live next to its producer.

Public API is unchanged: sh/__init__.py still re-exports CompletedCommand.
