+++ src/nclutils/sh/__init__.py
@@ -1,13 +1,12 @@
 """Run shell commands."""
 
 from .errors import (
-    CompletedCommand,
     ShellCommandError,
     ShellCommandFailedError,
     ShellCommandNotFoundError,
     ShellCommandTimeoutError,
 )
-from .shell_command import run_command, run_interactive, which
+from .shell_command import CompletedCommand, run_command, run_interactive, which
 
 __all__ = [
     "CompletedCommand",
