+++ src/nclutils/sh/errors.py
@@ -2,51 +2,12 @@
 
 from __future__ import annotations
 
-import shlex
-from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
-    from pathlib import Path
 
-
-@dataclass(frozen=True, slots=True)
-class CompletedCommand:
-    """The full record of a finished subprocess invocation.
-
-    Returned from :func:`run_command` on success and exposed as ``.result`` on
-    :class:`ShellCommandFailedError` / :class:`ShellCommandTimeoutError`.
-    Trailing newlines on ``stdout`` and ``stderr`` are stripped; embedded
-    newlines between lines are preserved.
-    """
-
-    argv: tuple[str, ...]
-    returncode: int
-    stdout: str
-    stderr: str
-    duration: float
-    cwd: Path | None
-
-    @property
-    def ok(self) -> bool:
-        """Return True when the command exited with status 0."""
-        return self.returncode == 0
-
-    @property
-    def command_line(self) -> str:
-        """Return ``argv`` rendered with :func:`shlex.join` so it can be pasted into a shell."""
-        return shlex.join(self.argv)
-
-    @property
-    def stdout_lines(self) -> list[str]:
-        """Return ``stdout`` split into lines. Recomputed on each access; bind locally to iterate twice."""
-        return self.stdout.splitlines()
-
-    @property
-    def stderr_lines(self) -> list[str]:
-        """Return ``stderr`` split into lines. Recomputed on each access; bind locally to iterate twice."""
-        return self.stderr.splitlines()
+    from .shell_command import CompletedCommand
 
 
 class ShellCommandError(Exception):
